+++ packages/create-pantheon-decoupled-kit/esbuild.js
@@ -8,9 +8,12 @@ const buildOptions = {
 	entryPoints: [
 		'./src/bin.ts',
 		// compile tagged templates to js
-		...(await glob('./src/templates/**/*.{css,jsx,tsx,ts,js,json,env.*}.ts', {
-			dot: true,
-		})),
+		...(await glob(
+			'./src/templates/**/*.{css,jsx,tsx,ts,js,json,md,env.*}.ts',
+			{
+				dot: true,
+			},
+		)),
 	],
 	bundle: true,
 	platform: 'node',