+++ packages/create-pantheon-decoupled-kit/src/templates/next-wp/README.md.ts
@@ -1,0 +1,38 @@
+import type { TemplateFn } from '@cli/types';
+import {
+	commands,
+	createPantheonDecoupledKitNpmPackage,
+	gettingStarted,
+	graphqlPostRequest,
+	healthCheck,
+	lando,
+	readmeHeader,
+	wordpressKitNpmPackage,
+} from '@partials/shared/readme/sharedReadme';
+
+const starterName = 'Next WordPress';
+
+const md: TemplateFn = () => /* md */ `${readmeHeader(starterName)}
+
+${gettingStarted(starterName)}
+
+## \`@pantheon-systems\` npm Packages
+
+${wordpressKitNpmPackage()}
+
+${createPantheonDecoupledKitNpmPackage()}
+
+${healthCheck('next-wp')}
+
+## Local development
+
+${lando('wordpress')}
+
+${commands(false)}
+
+## Miscellaneous
+
+${graphqlPostRequest()}
+`;
+
+export default md;