+++ packages/create-pantheon-decoupled-kit/src/utils/constants.ts
@@ -3,4 +3,4 @@
  * `env\.[a-z.]*` - matches `.env.*.ts` such as `.env.development.local.ts`
  */
 export const TAGGED_TEMPLATE_REGEX =
-	/(\.(css|jsx|tsx|ts|js|json|env\.[a-z.]*)\.(js|ts))/;
+	/(\.(css|jsx|tsx|ts|js|json|md|env\.[a-z.]*)\.(js|ts))/;