+++ packages/create-pantheon-decoupled-kit/CHANGELOG.md
@@ -1,5 +1,17 @@
 # create-pantheon-decoupled-kit
 
+## 0.10.2-canary.1
+
+### Patch Changes
+
+- 793e6268: Fix logo alignment in READMEs
+
+## 0.10.2-canary.0
+
+### Patch Changes
+
+- 6b616fb9: [next-wp][next-drupal][gatsby-wp] Update READMEs
+
 ## 0.10.1
 
 ### Patch Changes